scaffolding for nearest neighbour slicing backend.
